+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `unwindRepeatableSets` parameter to form store
+
 ## [1.1.1] - 2022-05-31
 
 ### Changed